add update create destroy controllers: game_buildings
