+++ Katan/app/assets/javascripts/game_buildings.coffee
