+++ Katan/app/assets/stylesheets/game_buildings.scss
