--- Katan/app/helpers/buildings_helper.rb
@@ -1,2 +1,0 @@
-module BuildingsHelper
-end
