+++ Katan/app/helpers/game_buildings_helper.rb
@@ -1,0 +1,2 @@
+module GameBuildingsHelper
+end
