--- Katan/app/views/buildings/index.html.erb
@@ -1,2 +1,0 @@
-<h1>Buildings#index</h1>
-<p>Find me in app/views/buildings/index.html.erb</p>
