+++ Katan/app/views/game_buildings/destroy.js.erb
@@ -1,0 +1,1 @@
+alert("<%= @result %>");
