+++ Katan/app/views/game_buildings/new.html.erb
