+++ Katan/app/views/game_buildings/show.html.erb
